Fastq dump rules now use parallel-fastq-dump.
Added parallel-fastq-dump to environment.yaml and rewrote both fastq dump rules (with and without prefetch) to use parallel-fastq-dump for multithreading.

I compared the output of the new rule against the old version (both non-prefetch). The gzipped files are different, but I confirmed the with md5 that the unzipped fastq are the same. I suspect parallel-fastq-dump compresses differently.