Set upper limit on x-axis cutoff in kmer_count_histogram.R.
Ok so the hardcoded cutoff at 150 was perfect for maize, perfect for potato, but too big for arabidopsis and too big for aliivibrio.

Now my clever 0.997 cumulative fraction cutoff was fantastic for aliivibrio but wayyy too big for maize. Setting the cumulative fraction to 0.95 was pretty nice for maize but obviously too small for aliivibrio.

So, new strategy: set x-axis cutoff at 150, unless the 0.997 cumulative fraction is smaller, in which case use that instead.