In kmer_count_histogram.R, improve logic of kmer peak finding.
Before, I found the peak by ignoring x-values to the left of 10 and finding the global maximum to the right. This strategy assumes the local minimum between the error peak and coverage peak will be around 10, and fails when that is not the case.

My new strategy is to approximate the first derivative in order to find the global maximum of the slope; this will reliably be at the inflection point on the left side of the coverage peak. Ignoring values to the left will effectively exclude the error peak, and the global maximum to the right will be the coverage peak. Woooo calculus!