After every rule using pigz, test gzipped output before continuing.
(I was having frequent failures with pigz complaining about crc32 mismatches. Not a bug in my code thankfully, my pigz install was broken. But this is a handy feature nonetheless.)

With this edit, every rule involving pigz will test the output using pigz -t. If the output is bad, pigz -t returns 1 so Snakemake will know the job was unsuccessful.