In kmer_score_histogram.R, x-axis cutoff is now 3 * most common k-mer score.
No longer hardcoded at 3000! That worked well for maize but not for smaller genomes.

I tried so many ideas to try to find when the curve becomes flat and/or approaches zero, but everything got so complicated because this data is much noisier than the k-mer counts and the tails look completely different between my large and small testing genomes (maize and aliivibrio fischeri). So this is my "worse is better" solution.