Made kmer_score_histogram.R about 60x faster for large genomes by filtering dataframe before attempting to plot.
I like the behavior of ggplot's coord_cartesian to zoom without clipping, but the flip side is that it wastes time by rendering observations that will fall outside the visible area. It took 10 minutes to plot the maize kmer score histogram because the dataframe had over 600,000 rows, even though the plot was cut off at just over 3000.

To reduce the plot time, I decided to filter the dataframe before passing to ggplot. I previously set the x-axis cutoff at the peak score * 3, so I set the filtering cutoff at the peak score * 6 to be conservative. Now the script plots the maize histogram in about 10 seconds.